requirements.txt
absl-py==2.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.19
etils==1.7.0
fonttools==4.51.0
fsspec==2024.3.1
glfw==2.7.0
gym==0.26.2
gym-notices==0.0.8
idna==3.7
importlib_resources==6.4.0
kiwisolver==1.4.5
labmaze==1.0.6
lmdb==1.5.1
lxml==5.2.2
matplotlib==3.8.4
mujoco==3.1.5
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
protobuf==5.26.1
PyOpenGL==3.1.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
scipy==1.13.0
six==1.16.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
zipp==3.18.1
